Day 9 - Support for large numbers

+++ src/main/java/intcode/EarlyIntcodeComputer.java
@@ -2,8 +2,8 @@ package intcode;
 
 public class EarlyIntcodeComputer {
 
-    public static int runNounVerb(int[] instructions, int noun, int verb) {
-        int[]  program = instructions.clone();
+    public static long runNounVerb(long[] instructions, long noun, long verb) {
+        long[]  program = instructions.clone();
         program[1] = noun;
         program[2] = verb;
 
@@ -13,7 +13,7 @@ public class EarlyIntcodeComputer {
         return computer.getProgram()[0];
     }
 
-    public static int runIO(int[] instructions, int input) {
+    public static long runIO(long[] instructions, long input) {
         IntcodeComputer computer = new IntcodeComputer(instructions);
         computer.run(input);
         return computer.getLastOutput();
